Move some test helpers out into a sperate file.
